Changed organization of load and dump
